move ReadWriteSpan from ddtrace to tracer package
