+++ ddtrace/example_test.go
@@ -11,7 +11,6 @@ import (
 
 	opentracing "github.com/opentracing/opentracing-go"
 
-	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/mocktracer"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/opentracer"
@@ -76,31 +75,4 @@ func Example_mocking() {
 	if spans[0].OperationName() != "test.span" {
 		// fail
 	}
-}
-
-// The code below illustrates how to set up a Post Processor in order to drop and/or modify traces.
-func Example_processor() {
-	// This processor will drop traces that do not contain an error, db span or client http request
-	// to endpoint GET /api/v1. In the case there is a http request to endpoint GET /api/v1, it will add
-	// a span tag to the local root span.
-	tracer.Start(tracer.WithPostProcessor(func(spans []ddtrace.ReadWriteSpan) (dropTrace bool) {
-		for _, s := range spans {
-			// trace contains an error which isn't "specific error".
-			if s.IsError() && s.Tag("error.message") != "specific error" {
-				return false
-			}
-			// trace contains a db request.
-			if s.Tag("span.type") == "db" {
-				return false
-			}
-			// trace contains a http request to endpoint GET /api/v1.
-			if s.Tag("service.name") == "service-a-http-client" && s.Tag("resource.name") == "GET /api/v1" {
-				// set tag on local root span.
-				spans[0].SetTag("calls.external.service", "service-b-api")
-				return false
-			}
-		}
-		return true
-	}))
-	defer tracer.Stop()
 }
