+++ ddtrace/tracer/example_test.go
@@ -51,3 +51,30 @@ func doSomething(ctx context.Context) (err error) {
 	}
 	return ctx.Err()
 }
+
+// The code below illustrates how to set up a Post Processor in order to drop and/or modify traces.
+func Example_processor() {
+	// This processor will drop traces that do not contain an error, db span or client http request
+	// to endpoint GET /api/v1. In the case there is a http request to endpoint GET /api/v1, it will add
+	// a span tag to the local root span.
+	tracer.Start(tracer.WithPostProcessor(func(spans []tracer.ReadWriteSpan) (dropTrace bool) {
+		for _, s := range spans {
+			// trace contains an error which isn't "specific error".
+			if s.IsError() && s.Tag("error.message") != "specific error" {
+				return false
+			}
+			// trace contains a db request.
+			if s.Tag("span.type") == "db" {
+				return false
+			}
+			// trace contains a http request to endpoint GET /api/v1.
+			if s.Tag("service.name") == "service-a-http-client" && s.Tag("resource.name") == "GET /api/v1" {
+				// set tag on local root span.
+				spans[0].SetTag("calls.external.service", "service-b-api")
+				return false
+			}
+		}
+		return true
+	}))
+	defer tracer.Stop()
+}
